Design-matrix columns UInt32 (was UInt16: wrapped above 65,535 matched precursors per scan)
timsTOF MS1 slices at 50 ng exceed the cap at the scout's +/-100 ppm; Hs.colptr[col + 1] wrapped to index 0.
Widened id_to_col / col / finalize_column! / SparsePrecMap / Score! signatures; regression test with 70,000
precursors in one scan (fails with BoundsError on the old code).

+++ src/Routines/SearchDIA/CommonSearchUtils/deconvolutionArrayUtils.jl
@@ -14,7 +14,7 @@ into `weights` (indexed by column number) for the current scan's active
 precursors.
 """
 function initialize_weights!(
-    id_to_col::AbstractPrecursorMap{UInt16},
+    id_to_col::AbstractPrecursorMap{UInt32},
     weights::Vector{Float32},
     precursor_weights::AbstractPrecursorMap{Float32}
 )
@@ -30,7 +30,7 @@ Copy solved weights back from `weights` (column-indexed) into
 `precursor_weights` (precursor-id-keyed) so subsequent scans can warm-start.
 """
 function update_precursor_weights!(
-    id_to_col::AbstractPrecursorMap{UInt16},
+    id_to_col::AbstractPrecursorMap{UInt32},
     weights::Vector{Float32},
     precursor_weights::AbstractPrecursorMap{Float32}
 )
